Game description
This Simple Game Programme is written in Python language. This Simple Game was created using Pygame. Pygame is a library of code that allows one to create simple games using Python. The purpose of this project was to adapt to working with code libraries. This Simple Game has One ‘player’ object, Three ‘enemy’ objects and One ‘prize’ object.